requirements.txt
-i https://pypi.org/simple
astroid==2.4.1
click==7.1.2
dnspython==1.16.0
email-validator==1.1.0
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pydantic==1.5.1
pydash==4.7.6
pyjwt==1.7.1
pylint==2.5.2
six==1.14.0
starlette==0.13.2
toml==0.10.1
uvicorn==0.11.5
uvloop==0.14.0 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1
wrapt==1.12.1
